+++ SkipList/SkipList/HashTable.h
@@ -10,11 +10,11 @@
 
 typedef struct HashTable {
 	int size;
-	int ratio;
+	float ratio;
 	LinkList** table;
 } HashTable;
 
-HashTable* HT_init(int size, int ratio);
+HashTable* HT_init(int size, float ratio);
 void HT_free(HashTable** list);
 LinkList* HT_Search(HashTable* list, int key);
 void HT_Insert(HashTable** list, int key, int value);
